move HttpBody to cli/http_util.rs
